backend/requirements.txt
# Backend Django - Dependencias de producción para Render
# TFG Idealista - Álvaro Carrera

# === DJANGO CORE ===
Django==4.2.7
gunicorn==21.2.0

# === DATABASE ===
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# === API & CORS ===
djangorestframework==3.14.0
django-cors-headers==4.3.1

# === STATIC FILES ===
whitenoise==6.6.0

# === CONFIGURATION ===
python-decouple==3.8
python-dotenv==1.0.0

# === DATA SCIENCE & ML ===
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4

# === VISUALIZATION ===
matplotlib==3.8.2
seaborn==0.13.0
folium==0.15.0
plotly==5.17.0

# === UTILITIES ===
requests==2.31.0